requirements.txt
# --- core app ---
streamlit==1.37.1
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
feedparser==6.0.11
python-dateutil==2.9.0.post0        # <-- added for date parsing (time windows)

# --- RAG / embeddings ---
datasets==2.20.0
sentence-transformers==2.6.1
faiss-cpu==1.13.0
torch==2.6.0

# --- LLMs (Gemini via LangChain) ---
google-generativeai==0.8.3
langchain-google-genai==2.0.5
langchain-core==0.3.19              # <- bumped to satisfy langchain-google-genai

# --- API server (optional) ---
fastapi==0.115.2
uvicorn[standard]==0.30.6

# --- compatibility pins ---
protobuf==5.29.2
grpcio==1.64.1
urllib3==2.2.3                     # <-- added to support Retry & HTTPAdapter safely
numpy<2.0.0,>=1.26.4